ShopStatus/requirements.txt
pillow
djangorestframework-simplejwt
django-filter
